mcp-servers/google-workspace-mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-workspace-mcp"
version = "0.1.0"
description = "MCP server for Google Workspace (Calendar, Tasks, Gmail, Sheets)"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "dateparser",
    "beautifulsoup4",
    "pdfplumber",
    "fastapi",
    "uvicorn",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
